+++ CHANGELOG.md
@@ -4,6 +4,32 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+- hardis:work:save: Do not git add manifest files when they have not been updated
+- Select type of org to connect: enhance label
+- Multi-Select default to 9999 items displayed
+- Display tips about deployment failures when they happen
+- Allow to define a file `manifest/deploymentPlan.json` to split the deployment into separate package.xml files
+
+Example:
+
+```json
+{
+  "packages": [
+    {
+      "label": "SharingRulesAccount",
+      "packageXmlFile": "splits/packageXmlSharingRulesAccount.xml",
+      "order": 10,
+      "waitAfter": 60
+    },
+    {
+      "label": "SharingRulesVisit__c",
+      "packageXmlFile": "splits/packageXmlSharingRulesAccountVisit__c.xml",
+      "order": 10
+    }
+  ]
+}
+```
+
 ## [2.2.1] 2021-03-23
 
 - QuickFix 2.2.1